Move constants to top-level module
