+++ scfw/configure/interactive.py
@@ -9,7 +9,7 @@ from typing import Optional
 
 import inquirer  # type: ignore
 
-from scfw.configure.constants import DD_API_KEY_VAR
+from scfw.constants import DD_API_KEY_VAR
 from scfw.logger import FirewallAction
 
 GREETING = (
