+++ scfw/constants.py
